requirements.txt
chromadb==0.6.3
langchain_chroma==0.2.3
langchain_core==0.3.52
langchain_openai==0.3.13
pulp==3.1.1
python-dotenv==1.1.0
streamlit==1.44.1